pyproject.toml
[tool.poetry]
name = "python-etl-data-quality"
version = "0.1.0"
description = ""
authors = ["Robinson <robinson.lovatto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.5"
mkdocs = "^1.5.3"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-material = "^9.5.13"
mkdocstrings = {extras = ["python"], version = "^0.24.1"}
taskipy = "^1.12.2"
isort = "^5.13.2"
black = "^24.2.0"
pytest = "^8.1.1"
sqlalchemy = "^2.0.28"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
psycopg2 = "^2.9.9"
pandera = {extras = ["io"], version = "^0.18.3"}
duckdb = "^0.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = { cmd = 'powershell -Command "Get-NetTCPConnection -LocalPort 8501 -ErrorAction SilentlyContinue | Select-Object -ExpandProperty OwningProcess | ForEach-Object {Stop-Process -Id $_ -Force}" '}
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"